fix(excuse): setMaxLength takes Int, use Kotlin javaObjectType
OptionData.setMaxLength is declared with Int; passing Long failed
compile. Also swap java.lang.Long::class.java to Long::class.javaObjectType
for the count-query result type to silence the
"This class is not recommended for use in Kotlin" warning.

+++ discord-bot/src/main/kotlin/bot/toby/command/commands/misc/ExcuseCommand.kt
@@ -31,7 +31,7 @@ class ExcuseCommand @Autowired constructor(
         SubcommandData(SUBMIT, "Submit a new excuse for approval (max 200 characters).")
             .addOptions(
                 OptionData(OptionType.STRING, OPT_TEXT, "The excuse text", true)
-                    .setMaxLength(MAX_EXCUSE_LENGTH.toLong()),
+                    .setMaxLength(MAX_EXCUSE_LENGTH),
                 OptionData(
                     OptionType.USER, OPT_AUTHOR,
                     "Attribute the excuse to this user instead of yourself", false
